requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
neo4j==5.14.0

# PDF Processing
pdfplumber==0.10.3
pytesseract==0.3.10
lxml==4.9.3

# NLP
spacy==3.7.2

# Translation & Validation
deepl==1.16.0
python-Levenshtein==0.23.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
mutmut==2.4.4
pact-python==2.0.1

# Development
python-dotenv==1.0.0
